+++ src/XUnity.AutoTranslator.Plugin.Core/UI/XuaWindow.cs
@@ -10,7 +10,7 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
    internal class XuaWindow
    {
       private const int WindowId = 5464332;
-      private const float WindowHeight = 520;
+      private const float WindowHeight = 540;
       private const float WindowWidth = 320;
 
       private Rect _windowRect = new Rect( 20, 20, WindowWidth, WindowHeight );
@@ -75,7 +75,7 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
 
             // GROUP
             var toggles = _viewModel.Toggles;
-            var groupHeight = ( GUIUtil.RowHeight * toggles.Count ) + ( GUIUtil.ComponentSpacing * ( toggles.Count ) ) - GUIUtil.HalfComponentSpacing;
+            var groupHeight = ( GUIUtil.RowHeight * toggles.Count ) + ( GUIUtil.HalfComponentSpacing * toggles.Count ) - GUIUtil.HalfComponentSpacing;
             GUI.Box( GUIUtil.R( GUIUtil.HalfComponentSpacing, posy, WindowWidth - GUIUtil.ComponentSpacing, groupHeight ), "" );
 
             foreach( var vm in toggles )
@@ -86,8 +86,9 @@ namespace XUnity.AutoTranslator.Plugin.Core.UI
                {
                   vm.OnToggled();
                }
-               posy += GUIUtil.RowHeight + GUIUtil.ComponentSpacing;
+               posy += GUIUtil.RowHeight;
             }
+            posy += GUIUtil.ComponentSpacing;
 
             var commandButtons = _viewModel.CommandButtons;
             const int buttonsPerRow = 3;